Update Advanced ToDoList
