+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/AdvancedToDoList.Android/MainActivity.cs
@@ -1,10 +1,6 @@
 ﻿using Android.App;
 using Android.Content.PM;
-using Avalonia;
 using Avalonia.Android;
-using AdvancedToDoList.Android.Services;
-using AdvancedToDoList.Services;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace AdvancedToDoList.Android;
 
@@ -14,18 +10,6 @@ namespace AdvancedToDoList.Android;
     Icon = "@drawable/icon",
     MainLauncher = true,
     ConfigurationChanges = ConfigChanges.Orientation | ConfigChanges.ScreenSize | ConfigChanges.UiMode)]
-public class MainActivity : AvaloniaMainActivity<App>
+public class MainActivity : AvaloniaMainActivity
 {
-    protected override AppBuilder CustomizeAppBuilder(AppBuilder builder)
-    {
-        // Register the Android service
-        var services = new ServiceCollection();
-        services.AddSingleton<IDatabaseService>(new AndroidDbService());
-        services.AddSingleton<ISettingsStorageService>(new DefaultSettingsStorageService());
-        
-        App.RegisterAppServices(services);
-
-        return base.CustomizeAppBuilder(builder)
-            .WithInterFont();
-    }
 }
