+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/AdvancedToDoList.Desktop/AdvancedToDoList.Desktop.csproj
@@ -17,11 +17,7 @@
   
   <ItemGroup>
     <PackageReference Include="Avalonia.Desktop" />
-    <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-    <PackageReference Include="Avalonia.Diagnostics" >
-      <IncludeAssets Condition="'$(Configuration)' != 'Debug'">None</IncludeAssets>
-      <PrivateAssets Condition="'$(Configuration)' != 'Debug'">All</PrivateAssets>
-    </PackageReference>
+    <PackageReference Include="AvaloniaUI.DiagnosticsSupport" />
     
     <!-- Logging -->
     <PackageReference Include="Serilog"/>
