+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/AdvancedToDoList.Desktop/Program.cs
@@ -51,6 +51,7 @@ sealed class Program
         => AppBuilder.Configure<App>()
             .UsePlatformDetect()
             .WithInterFont()
+            .WithDeveloperTools()
             .LogToTrace();
 
     private static void ConfigureLogging()
