+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/AdvancedToDoList/Views/MainView.axaml
@@ -5,7 +5,6 @@
              xmlns:vm="using:AdvancedToDoList.ViewModels"
              xmlns:views="clr-namespace:AdvancedToDoList.Views"
              xmlns:controls="clr-namespace:SharedControls.Controls;assembly=SharedControls"
-             xmlns:sys="clr-namespace:System;assembly=System.Runtime"
              mc:Ignorable="d" d:DesignWidth="690" d:DesignHeight="450"
              x:Class="AdvancedToDoList.Views.MainView"
              x:Name="Root"
