+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/SharedControls/Controls/Dialogs/DialogHost.cs
@@ -72,7 +72,7 @@ public class DialogHost : ContentControl
     {
         base.OnAttachedToVisualTree(e);
     
-        if (e.Root is Control root)
+        if (TopLevel.GetTopLevel(this) is { } root)
         {
             // OverlayLayer is a Canvas, so we won't get a signal to resize if the window
             // bounds change. Subscribe to force update
