+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/SharedControls/Controls/Dialogs/DialogResult.cs
@@ -1,3 +1,5 @@
+using SharedControls.Services;
+
 namespace SharedControls.Controls;
 
 /// <summary>
