+++ src/Avalonia.Samples/CompleteApps/AdvancedToDoList/SharedControlsTests/SharedControlsTests.csproj
@@ -9,7 +9,7 @@
     <ItemGroup>
         <PackageReference Include="Avalonia" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" />
-        <PackageReference Include="XUnit" />
+        <PackageReference Include="XUnit.v3" />
         <PackageReference Include="Avalonia.Headless.XUnit" />
     </ItemGroup>
